[MOD] Improved Hamburger Menu animation and routing
